tor: move AddOnion to its own file
